Some review comments fixes.
