+++ dozer-spring-boot-auto-configuration/dozer-spring-boot-starter/pom.xml
@@ -19,8 +19,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>dozer-spring-boot-auto-configuration</artifactId>
         <groupId>com.github.dozermapper</groupId>
+        <artifactId>dozer-spring-boot-auto-configuration</artifactId>
         <version>6.2.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
